Add externs and ^:export to improve function under advanced comp.
Note I said "improve," not "fix." As noted in the project.clj
comment in the :min profile, under :advanced optimizations it
doesn't seem to export sheater.core.init properly, even though
it does seem to call the goog.exportSymbol. Not sure what's going
on, but at least the gapi stuff seems to be functional. Hopefully
if we can fix (init) not being called, everything will "just work."

With :simple optimizations, app.js is a whopping 9mb; under advanced,
it's "only" 1.7mb so we definitely need that. Hopefully the huge
gains aren't just because it's not actually including sheater.core :/