Change :value to inflate to a callable.
Fixes #8

This callable is, well, *called* in the render function to
retrieve the actual value. We need to do this for form-2
components, so they connect to a reaction and can get updated
when their source data changes.

Perhaps slightly fancier would be to provide a (->value)
function instead of directly pulling the :value key, then
only making it callable when strictly necessary, and ->value
handles that case for us. That'd be cool, but it seems a bit
too magic right now.