Fix support for negative numbers
Also, refactored all the duplicate implementations of number inputs
to just use our custom input, and updated that to support both
auto and manual modes.

The big problem with numbers was a bug in ReactJS. Reagent hasn't
released the new version that uses the bugfix ReactJS release, so
I manually updated the dependencies and that seems to work for now.
Also, I had to update the number regex to accept negative numbers.

Derp.